twitter_persona/requirements/requirements.txt
tweepy>=4.14.0
anthropic>=0.18.1
python-dotenv>=1.0.0
loguru>=0.7.2 